pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-payments-chile"
version = "2024.12.2a"
description = "Pasarelas de pago chilenas para django-payments"
readme = "README.md"
authors = [{ "name" = "Mario Hernandez", "email" = "mariofix@proton.me" }]
license = { text = "MIT License" }
requires-python = ">=3.9,<4.0"

keywords = [
    "flow",
    "khipu",
    "kushki",
    "transbank",
    "payku",
    "pagofacil",
    "webpay",
    "onepay",
    "pagos",
    "django",
    "payment",
    "django-payments",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: Django",
    "Operating System :: OS Independent",
]

dependencies = ["django-payments>=2.0, <4.0", "requests", "pyflowcl"]

[project.urls]
Homepage = "https://mariofix.github.io/django-payments-chile/"
Repository = "https://github.com/mariofix/django-payments-chile"
Documentation = "https://mariofix.github.io/django-payments-chile/"


[project.optional-dependencies]
webpay = ["transbank-sdk>=5.0.0"]
oneclick = ["transbank-sdk>=5.0.0"]
todos = ["transbank-sdk>=5.0.0"]
dev = [
    "pre-commit",
    "black",
    "isort",
    "coverage",
    "pytest",
    "faker",
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
    "mkdocs-git-revision-date-localized-plugin",
    "pillow",
    "cairosvg",
    "lxml",
    "pytest-cov",
    "pytest-django",
    "pytest-mock",
    "dj-database-url",
    "factory-boy",
    "tox",
    "tox-gh-actions",
]


[tool.setuptools.packages.find]
include = ["django_payments_chile", "tests"]

[tool.black]
line-length = 119
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 5
py_version = 39
src_paths = ["django_payments_chile", "tests"]


[tool.coverage.report]
exclude_lines = ["if TYPE_CHECKING:"]

[tool.django-stubs]
django_settings_module = "tests.django_settings"

[tool.pytest.ini_options]
addopts = [
    "--cov=django_payments_chile",
    "--cov-report=term-missing:skip-covered",
    "--no-cov-on-fail",
    "--color=yes",
]
testpaths = "tests"
DJANGO_SETTINGS_MODULE = "tests.django_settings"
pythonpath = "."
